Title: 달에서 무게 구하기, Time: 99ms, Memory: 9MB, Status: Passed - CodeTree

+++ 250311/달에서 무게 구하기/weight-on-the-moon.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         // Please write your code here.
-        double weight = 13;
+        int weight = 13;
         double radio = 0.165;
         double result = weight * radio;
